feat: replacerをobjectに切り出し
