+++ src/main/kotlin/com/jaoafa/vcspeaker/voicetext/textreplacers/ChannelMentionReplacer.kt
@@ -1,0 +1,10 @@
+package com.jaoafa.vcspeaker.voicetext.textreplacers
+
+import dev.kord.common.entity.Snowflake
+
+object ChannelMentionReplacer : BaseReplacer {
+    override suspend fun replace(text: String, guildId: Snowflake): String =
+        replaceMentionable(text, Regex("<#(\\d+)>")) { kord, id ->
+            kord.getChannel(id)?.data?.name?.value ?: "不明なチャンネル"
+        }
+}
