+++ src/main/kotlin/com/jaoafa/vcspeaker/voicetext/textreplacers/EmojiReplacer.kt
@@ -1,0 +1,11 @@
+package com.jaoafa.vcspeaker.voicetext.textreplacers
+
+import com.jaoafa.vcspeaker.stores.AliasType
+import dev.kord.common.entity.Snowflake
+
+object EmojiReplacer : BaseReplacer {
+    override suspend fun replace(text: String, guildId: Snowflake) =
+        replaceText(text, guildId, AliasType.Emoji) { alias, replacedText ->
+            replacedText.replace(alias.from, alias.to)
+        }
+}
